+++ REPLesent.scala
@@ -300,7 +300,7 @@ case class REPLesent(
 
     def nextBuild: Option[Build] = select(build = buildCursor + 1) orElse jump(1)
 
-    def refreshBuild: Option[Build] = select(build = buildCursor)
+    def redrawBuild: Option[Build] = select(build = buildCursor)
 
     def previousBuild: Option[Build] = select(build = buildCursor - 1) orElse {
       jump(-1) flatMap { _ =>
@@ -328,6 +328,7 @@ case class REPLesent(
   private val helpMessage = """Usage:
     |  next          n      >     go to next build/slide
     |  previous      p      <     go back to previous build/slide
+    |  redraw        z            redraw the current build/slide
     |  Next          N      >>    go to next slide
     |  Previous      P      <<    go back to previous slide
     |  i next        i n          advance i slides
@@ -499,7 +500,8 @@ case class REPLesent(
   def p: Unit = previous
   def < : Unit = previous
 
-  def z: Unit = show(deck.refreshBuild)
+  def redraw: Unit = show(deck.redrawBuild)
+  def z: Unit = redraw
 
   def Next: Unit = 1.next
   def N: Unit = Next